Move mnist loader to separate file. Split Simple and CNN example to separate files.
